move tests from chocon_test.go to proxy/handler_test.go
